refactor: remove code from index so its easier to test
will be injecting custom db logger etc
