+++ src/database.js
@@ -139,43 +139,3 @@ export default class Database {
     }
   }
 }
-
-// Export legacy functions for backward compatibility
-export function getSession(
-  uri = "bolt://localhost:7687",
-  username = "",
-  password = ""
-) {
-  const db = new Database(uri, username, password);
-  return db.getSession();
-}
-
-export async function deleteNodes() {
-  const db = new Database();
-  return await db.deleteNodes();
-}
-
-export async function insertFileNode(file) {
-  const db = new Database();
-  return await db.insertFileNode(file);
-}
-
-export async function insertMethodNode(method) {
-  const db = new Database();
-  return await db.insertMethodNode(method);
-}
-
-export async function insertClassNode(klass) {
-  const db = new Database();
-  return await db.insertClassNode(klass);
-}
-
-export async function createRelationship(fromId, toId, type) {
-  const db = new Database();
-  return await db.createRelationship(fromId, toId, type);
-}
-
-export async function findAllMethods() {
-  const db = new Database();
-  return await db.findAllMethods();
-}
