+++ src/workspace-crawler.js
@@ -1,0 +1,88 @@
+import { findSupportedFiles } from "./file_utils.js";
+import LspClient from "./lsp-client.js";
+import FileCrawler from "./file-crawler.js";
+export default class WorkspaceCrawler {
+  constructor({ rootDir, logger, db }) {
+    this.rootDir = rootDir;
+    this.logger = logger;
+    this.db = db;
+  }
+
+  async crawl(files) {
+    const filesByLanguageHash = findSupportedFiles(this.rootDir);
+
+    for (const language in filesByLanguageHash) {
+      await this.processLanguage(
+        language,
+        filesByLanguageHash[language],
+        this.logger,
+        this.db
+      );
+    }
+  }
+
+  async processLanguage(language, files, logger, db) {
+    const lspClient = new LspClient({
+      language,
+      logger,
+      rootPath: process.argv[2],
+    });
+
+    await lspClient.connect();
+    for (const file of files) {
+      await lspClient.notifyFileOpen(file.uri, file.source);
+    }
+
+    for (const file of files) {
+      await new FileCrawler(file, this.logger, lspClient, this.db).crawl();
+    }
+
+    this.logger.info("Finding all inserted methods");
+    const { methodByFileMap, methods } = await this.db.findAllMethods(language);
+
+    this.logger.info("Finding all method references");
+    const methodsAndReferences = await Promise.all(
+      methods.map(async (method) => {
+        const result = await lspClient.findAllReferences(method, method.file);
+        const mappedReferences = result.map((reference) => {
+          /**
+           * this will break for child methods potentially
+           * function myFunc(arr) {
+           *  return arr.map(YourFunc)
+           * }
+           *  yourFunc reference call will match to both myFunc and YourFunc since we are
+           *  just comparing if it is within the range of the method, and we are not managing child
+           *  references
+           */
+          return (methodByFileMap[reference.uri] || []).find((method) => {
+            return (
+              reference.uri == method.file &&
+              reference.range.start.line > method.range.start.line &&
+              reference.range.start.line < method.range.end.line
+            );
+          });
+        });
+        return {
+          method,
+          references: result,
+          mappedReferences,
+        };
+      })
+    ).catch((err) => this.logger.error(err));
+
+    this.logger.info("Linking all method references in database");
+    for (let i = 0; i < methodsAndReferences.length; i++) {
+      const methodAndReference = methodsAndReferences[i];
+      for (const reference of methodAndReference.mappedReferences) {
+        if (reference) {
+          await this.db.createRelationship(
+            reference.id,
+            methodAndReference.method.id,
+            "CALLS"
+          );
+        }
+      }
+    }
+    lspClient.shutdown();
+  }
+}
